Removed Conditional Theme Inclusion
There was a check in config.php:
if ('ltr' === get_string('thisdirection', 'langconfig')) {
$THEME->sheets = array('moodle');
} else {
$THEME->sheets = array('moodle-rtl');
}
I am assuming this was originally part of some experiment, but it fails
because styles.php does not have access to the language setting of the
session. Normally leaving this code would be harmless but (a) it is
entirely redundant as is the large moodle-rtl.css it refers to, and (b)
for some reason it triggered on one of the systems I worked with, so I
removed it, together with the moodle-rtl css.